requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.3
openpyxl==3.1.2
plotly==5.18.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pillow==10.1.0
